Now encrypting string attributes.

+++ Employees/BasePlusCommissionEmployee.cs
@@ -35,9 +35,9 @@ namespace Employees
             }
         }
 
-        public override float CalcGrossPay()
+        public override decimal CalcGrossPay()
         {
-            return (float)(BaseSalary + (decimal) base.CalcGrossPay());
+            return BaseSalary + base.CalcGrossPay();
         }
     }
 }
